refactor: move polling errors into it's on module
