+++ src/services/challenge-polling-service.ts
@@ -1,5 +1,9 @@
 import { SIGN_ALGORITHM } from "@/constants/auth";
 import {
+  ChallengePollingNetworkError,
+  ChallengePollingServerError,
+} from "@/errors/challenge-polling-errors";
+import {
   PushRequest,
   PushRequestStatus,
   PushToken,
@@ -20,23 +24,6 @@ export interface TokenChallengePollingResult {
   tokenId: string;
   success: boolean;
   error?: Error;
-}
-
-export class ChallengePollingNetworkError extends Error {
-  constructor(cause: Error) {
-    super(cause.message, { cause });
-    this.name = "ChallengePollingNetworkError";
-  }
-}
-
-export class ChallengePollingServerError extends Error {
-  constructor(
-    public readonly status: number,
-    public readonly responseBody: string,
-  ) {
-    super(`Server returned ${status}: ${responseBody}`);
-    this.name = "ChallengePollingServerError";
-  }
 }
 
 interface ChallengeResponse {
